Renamed test file, and organised tests
